+++ TAS.Client.Config/Model/IngestDirectory.cs
@@ -14,13 +14,13 @@ namespace TAS.Client.Config.Model
         public IngestDirectory()
         {
             IsImport = true;
-            VideoBitrateRatio = 1.0M;
-            AudioBitrateRatio = 1.0M;
+            VideoBitrateRatio = 1;
+            AudioBitrateRatio = 1;
         }
         [DefaultValue(default(TAspectConversion))]
         public TAspectConversion AspectConversion { get; set; }
-        [DefaultValue(typeof(Decimal), "0")]
-        public decimal AudioVolume { get; set; }
+        [DefaultValue(typeof(double), "0")]
+        public double AudioVolume { get; set; }
         [DefaultValue(false)]
         public bool DeleteSource { get; set; }
         [DefaultValue(default(TIngestDirectoryKind))]
@@ -68,10 +68,10 @@ namespace TAS.Client.Config.Model
         public string[] Extensions { get; set; }
         public TVideoCodec VideoCodec { get; set; }
         public TAudioCodec AudioCodec { get; set; }
-        [DefaultValue(typeof(Decimal), "1")]
-        public decimal VideoBitrateRatio { get; set; }
-        [DefaultValue(typeof(Decimal), "1")]
-        public decimal AudioBitrateRatio { get; set; }
+        [DefaultValue(typeof(double), "1")]
+        public double VideoBitrateRatio { get; set; }
+        [DefaultValue(typeof(double), "1")]
+        public double AudioBitrateRatio { get; set; }
         [XmlArray(nameof(SubDirectories))]
         public IngestDirectory[] SubDirectoriesSerialized = new IngestDirectory[0];
         [XmlIgnore]
