+++ TAS.Common/Interfaces/IIngestOperation.cs
@@ -7,7 +7,7 @@ namespace TAS.Common.Interfaces
         TAspectConversion AspectConversion { get; set; }
         TAudioChannelMappingConversion AudioChannelMappingConversion { get; set; }
         TFieldOrder SourceFieldOrderEnforceConversion { get; set; }
-        decimal AudioVolume { get; set; }
+        double AudioVolume { get; set; }
         bool Trim { get; set; }
         bool LoudnessCheck { get; set; }
         TimeSpan StartTC { get; set; }
