+++ TAS.Common/Interfaces/IPreview.cs
@@ -4,7 +4,7 @@ namespace TAS.Common.Interfaces
 {
     public interface IPreview: INotifyPropertyChanged
     {
-        void PreviewLoad(IMedia media, long seek, long duration, long position, decimal audioLevel);
+        void PreviewLoad(IMedia media, long seek, long duration, long position, double audioLevel);
         IMedia PreviewMedia { get; }
         IPlayoutServerChannel PlayoutChannelPRV { get; }
         VideoFormatDescription FormatDescription { get; }
@@ -15,6 +15,6 @@ namespace TAS.Common.Interfaces
         bool PreviewIsPlaying { get; }
         long PreviewPosition { get; set; }
         long PreviewSeek { get; }
-        decimal PreviewAudioVolume { get; set; }
+        double PreviewAudioVolume { get; set; }
     }
 }
